tests/wpt/tests/fedcm/fedcm-button-and-other-account/fedcm-button-mode-basics.tentative.https.html

<!DOCTYPE html>
<title>Federated Credential Management API Active Mode basic tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_test,
        select_manifest,
        fedcm_get_and_select_first_account} from '../support/fedcm-helper.sub.js';

fedcm_test(async t => {
  let test_options = request_options_with_mediation_required();
  test_options.identity.mode = "active";
  await select_manifest(t, test_options);

  let result = navigator.credentials.get(test_options);
  return promise_rejects_dom(t, 'NetworkError', result);
}, "Test that the active mode without user activation will fail.");

fedcm_test(async t => {
  let test_options = request_options_with_mediation_required("manifest_with_rp_mode.json");
  test_options.identity.mode = "active";

  return test_driver.bless('initiate FedCM request', async function() {
      let cred = await fedcm_get_and_select_first_account(t, test_options);
      assert_equals(cred.token, "mode=active");
  });
}, "Test that the active mode succeeds with user activation.");

fedcm_test(async t => {
  let test_options = request_options_with_mediation_required("manifest_with_rp_mode.json");
  test_options.identity.mode = "active";
  test_options.mediation = "silent";

  return test_driver.bless('initiate FedCM request', async function() {
      let cred = fedcm_get_and_select_first_account(t, test_options);
      let rej = promise_rejects_dom(t, 'NotSupportedError', cred);
      await rej;
  });
}, "Test that the mediation:silent is not supported in active mode.");
</script>
